editoast: move map_to_core_work_schedule to WorkSchedule
